Phase 0: fix within_day_temporal split granularity, sweep guard (a), fix autoencoder
Partition (the split-strategy bug):
- Split within each (day, family) block, not each day. CICIDS runs attacks in
  sequential daily time blocks, so a per-day cut routed whole families to single
  partitions and re-created the day-split problem at finer granularity. Now every
  attack family lands in every partition with temporal ordering intact; all 13
  real CICIDS families are seen_both under within_day.
- Eval-family reporting is four-way: seen_both / seed_only (A4) / honeypot_only
  (decoy-only teaching, S0's key arm) / novel. Set explicitly via
  pool_withheld_families and seed_withheld_families config, not left to chance.
  Per-family TPR reported with trusted_eval row counts; families with <500 eval
  rows excluded from per-family reporting. arm_source column flags whether a
  non-seen_both arm was configured or is a split artifact.
- leak_warning is gated on the attack class only: near-identical benign flows
  across partitions are normal for real traffic, not a memorization leak. With
  this, day_split on real CICIDS is leak-free (attack NN p5 0.65) while
  within_day still leaks on DoS degeneracy (0.003).

Guard (a) grid sweep (experiments/phase0_grid_sweep.py, DECISIONS.md sec.11):
- Sweeps near_dup_grid in {off, 0.005, 0.01, 0.02, 0.05}, reports rows removed
  per class + RF/XGB metrics + the leak warning.
- CICIDS: guard is necessary (off -> 12.3k content-identical flows span
  seed_train/trusted_eval, partition invalid); model metrics are flat across the
  0.005-0.05 range despite removing 574k-1075k benign rows; the leak warning is
  set at every grid. The guard is not doing the attack's job. Default stays 0.02.
- Synthetic: completely flat.

Autoencoder (DECISIONS.md sec.8, sec.12): the normalizer exploded near-constant
CICIDS features (IQR exactly zero for flag counts). Floor the scale at
max(IQR, 0.1) and clip the transform to +-25. AUROC 0.59 -> ~0.89 on CICIDS,
~0.91 on synthetic; kept. At 1% FPR it still catches almost nothing (heavy
benign reconstruction-error tail) so S0/A1 use RF and XGBoost.

Memory: features stored float32 through the partition; row-content and near-dup
hashing are rolling hashes computed one column at a time; the disjointness
time-check no longer concatenates all partitions. Needed to fit CICIDS in ~2 GB.

Results regenerated for synthetic and CICIDS, both strategies; 62 tests pass.

+++ src/dloop/features/cleaning.py
@@ -186,7 +186,10 @@ def clean_flows(df: pd.DataFrame, *, source: str) -> tuple[pd.DataFrame, Cleanin
             np.arange(len(df)), unit="s"
         )
 
-    # Step 7: finalize schema.
+    # Step 7: finalize schema. Features stored as float32 — flow summary
+    # statistics do not need double precision, and CICIDS2017 is ~2.8M rows, so
+    # this halves the memory the partition and leakage guards work in.
+    df[feats] = df[feats].astype("float32")
     df[schema.BINARY_LABEL] = (df[schema.LABEL] != schema.BENIGN_LABEL).astype("int8")
     if schema.DAY not in df.columns:
         df[schema.DAY] = source
